refactor migration structure
