+++ .gitignore
@@ -5,7 +5,6 @@ _node_modules
 **/node_modules
 # Production
 /build
-/tdevMigrations
 migrateTdev.config.yaml
 # Generated files
 .docusaurus
