--- packages/tdev/material-sync/src/migrate_tdev/.gitignore
@@ -1,1 +1,0 @@
-migrations/
