+++ packages/tdev/material-sync/src/migrate_tdev/README.mdx
@@ -1,0 +1,7 @@
+---
+page_id: 35817c9b-3336-4fd4-9749-1fd973851568
+---
+
+# Migrate TDev
+
+Simples CLI tool, um mehrere Teaching-Dev Projekte zu migrieren.
