+++ packages/tdev/material-sync/src/migrate_tdev/helpers/loadMigrationRunners.ts
@@ -1,6 +1,6 @@
 import fs from 'node:fs/promises';
 import path from 'node:path';
-import { MIGRATION_PATH, MigrationRunner } from './constants';
+import { MIGRATION_PATH, MigrationRunner } from '../constants';
 
 export async function* loadMigrationRunners(): AsyncGenerator<MigrationRunner> {
     const migrationFiles = await fs.readdir(MIGRATION_PATH);
