+++ packages/tdev/material-sync/src/migrate_tdev/index.ts
@@ -3,16 +3,16 @@ import path from 'node:path';
 import Rsync from 'rsync';
 import { REPO_ROOT, PACKAGE_ROOT, pathExistsSync } from '../helpers.js';
 import { fileURLToPath } from 'node:url';
-import readOrCreateConfig from './readOrCreateConfig.js';
+import readOrCreateMigrationConfig from './helpers/readOrCreateMigrationConfig.js';
 import { MIGRATION_PATH } from './constants.js';
-import { loadMigrationRunners } from './loadMigrationRunners.js';
-import { gitEnsureClean } from './helpers.js';
+import { loadMigrationRunners } from './helpers/loadMigrationRunners.js';
+import { gitEnsureClean } from './helpers/actions.js';
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 process.chdir(REPO_ROOT);
 
 const main = async (): Promise<void> => {
-    const config = await readOrCreateConfig();
+    const config = await readOrCreateMigrationConfig();
     console.log('Migrating...', MIGRATION_PATH);
     for await (const runMigration of loadMigrationRunners()) {
         for (const tdevPage of config.tdevPages) {
