--- packages/tdev/material-sync/src/migrate_tdev/migrateTdev.default.yaml
@@ -1,1 +1,0 @@
-tdevPages: []
